Move getRecordFieldSelectors into a separate file
