+++ packages/graph/src/index.ts
@@ -5,3 +5,4 @@ export * from './layers/DataLayer';
 export * from './layers/AxisLayer';
 export * from './layers/HoverLayer';
 export * from './tooltip/Tooltip';
+export * from './utils/getRecordFieldSelectors';
